\name{m}
\alias{m}
\title{
  Just a quick alias for the function as.matrix
}
\description{
  Just a quick alias for the function as.matrix
}
\usage{
m(x, ...)
}
\arguments{
  \item{x}{an R object}
  \item{\dots}{additionnal arguments to \code{as.matrix}}
}
\value{
  a matrix
}
\seealso{
  \code{\link{as.matrix}}
}
\examples{
## No examples
}
\keyword{math}
