\name{oncfsMisc-package}
\alias{oncfsMisc-package}
\alias{oncfsMisc}
\docType{package}
\title{
  Various functions used to analyse datasets at the French Biodiversity organization.
}
\description{
  This package contains various functions first developed at the ONCFS (French
  wildlife management organization) -- now OFB (French organization for biodiversity).
}
\details{
\tabular{ll}{
Package: \tab oncfsMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-09\cr
License: \tab GNU GPL > 2\cr
}
}
\author{
  Clement Calenge \email{clement.calenge@ofb.gouv.fr}
}
\keyword{ package }
\examples{
## No examples
}
