\name{oncfsMisc-package}
\alias{oncfsMisc-package}
\alias{oncfsMisc}
\docType{package}
\title{
  Various functions used to analyse datasets at the ONCFS
}
\description{
  This package contains various functions used at the ONCFS (French
  wildlife management organization; \url{http://www.oncfs.gouv.fr/})
}
\details{
\tabular{ll}{
Package: \tab oncfsMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-09\cr
License: \tab GNU GPL > 2\cr
}
}
\author{
  Clement Calenge \email{clement.calenge@oncfs.gouv.fr}
}
\keyword{ package }
\examples{
## No examples
}
